requirements.txt
fastapi
uvicorn
pydantic
huggingface_hub
google-generativeai
Pillow
